Project migration to vite, typescript, postgresql
