--- README.md
@@ -1,202 +1,0 @@
-<a name="readme-top"></a>
-
-<!-- PROJECT SHIELDS -->
-<!--
-*** I'm using markdown "reference style" links for readability.
-*** Reference links are enclosed in brackets [ ] instead of parentheses ( ).
-*** See the bottom of this document for the declaration of the reference variables
-*** for contributors-url, forks-url, etc. This is an optional, concise syntax you may use.
-*** https://www.markdownguide.org/basic-syntax/#reference-style-links
--->
-
-[![MIT License][license-shield]][license-url]
-[![LinkedIn][linkedin-shield]][linkedin-url]
-
-<!-- PROJECT LOGO -->
-<br />
-<div align="center">
-  <a href="https://github.com/Asalz20/bible-verse-scripture-app">
-    <!-- <img src="./imgs/alex-logo.png" alt="Logo" width="120" height="80"> -->
-  </a>
-
-  <h3 align="center">Church of Bergen Scripture Study App</h3>
-
-  <p align="center">
-    A Bible Scripture study app used by the Church of Bergen to recite, study, and interpret scripture.
-    <!-- <br />
-    <a href="https://pectusplus-blog.vercel.app/">View Demo</a>
-    ·
-    <a href="https://github.com/Asalz20/pectusplus-blog/issues">Report Bug</a>
-  </p> -->
-</div>
-
-<!-- TABLE OF CONTENTS -->
-<details>
-  <summary>Table of Contents</summary>
-  <ol>
-    <li>
-      <a href="#about-the-project">About The Project</a>
-      <ul>
-        <li><a href="#built-with">Built With</a></li>
-      </ul>
-    </li>
-    <!-- <li>
-      <a href="#getting-started">Getting Started</a>
-      <ul>
-        <li><a href="#prerequisites">Prerequisites</a></li>
-        <li><a href="#installation">Installation</a></li>
-      </ul>
-    </li> -->
-    <li><a href="#roadmap">Roadmap</a></li>
-    <li><a href="#license">License</a></li>
-    <li><a href="#contact">Contact</a></li>
-    <li><a href="#acknowledgments">Acknowledgments</a></li>
-  </ol>
-</details>
-
-<!-- ABOUT THE PROJECT -->
-
-## About The Project
-
-<!-- [![Product Name Screen Shot][product-screenshot]](https://pectusplus-blog.vercel.app/) -->
-
-The Bible Verse Study App was created to aid the Church of Bergen in their teachings to their students.
-
-BVSA attempts to:
-
-- Keep education and information accessible for all users
-- Bring together a religious community for empowerment through scripture
-
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
-
-### Built With
-
-The following are major frameworks/libraries used to bootstrap this project.
-
-[![TypeScript][typescript-shield]][typescript-url] [![CSS][css-shield]][css-url] [![React][react.js]][react-url] [![Node][node-shield]][node-url] [![Express][express-shield]][express-url] [![PostgreSQL][postgres-shield]][postgres-url]
-
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
-
-<!-- GETTING STARTED -->
-
-<!-- ## Getting Started
-
-To get a local copy up and running follow these steps.
-
-### Prerequisites
-
-Prior to proceeding, ensure that npm has been installed. Use the following to check npm version
-
-```sh
-npm --v
-```
-
-If it has not yet been installed, follow the appropriate steps to do so.
-
-```sh
-npm install npm@latest -g
-```
-
-### Installation
-
-1. Clone the repo
-   ```sh
-   git clone https://github.com/Asalz20/pectusplus-blog.git
-   ```
-2. Install NPM packages at root, web, and studio directory
-   ```sh
-   npm install
-   ```
-3. In order to spin up your own sanity backend, head over to [Sanity.io](https://sanity.io/) and follow their Getting started documentation.
-
-4. Run the following in the web directory to start the development server
-
-```sh
-   npm start
-```
-
-<p align="right">(<a href="#readme-top">back to top</a>)</p> -->
-
-<!-- USAGE EXAMPLES -->
-
-<!-- ## Usage
-
-Use this space to show useful examples of how a project can be used. Additional screenshots, code examples and demos work well in this space. You may also link to more resources.
-
-_For more examples, please refer to the [Documentation](https://example.com)_
-
-<p align="right">(<a href="#readme-top">back to top</a>)</p> -->
-
-<!-- ROADMAP -->
-
-## Roadmap
-
-- [x] Allow users to save multiple cards
-- [ ] Ability to go through deck
-- [ ] Compare to cards to teacher interpretation
-
-See the [open issues](https://github.com/Asalz20/bible-verse-scripture-app/issues) for a full list of proposed features (and known issues).
-
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
-
-<!-- LICENSE -->
-
-## License
-
-Distributed under the MIT License. See `LICENSE.txt` for more information.
-
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
-
-<!-- CONTACT -->
-
-## Contact
-
-Alex Salazar - [@alexsalazarr\_](https://twitter.com/alexsalazarr_) - asalazar.dev@gmail.com
-
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
-
-<!-- ACKNOWLEDGMENTS -->
-
-## Acknowledgments
-
-Giving credit where credit is due! Thank you to:
-
-- [Img Shields](https://shields.io)
-- [GitHub Pages](https://pages.github.com)
-- [Sanity.io](https://sanity.io)
-- [React Icons](https://react-icons.github.io/react-icons/search)
-
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
-
-<!-- MARKDOWN LINKS & IMAGES -->
-<!-- https://www.markdownguide.org/basic-syntax/#reference-style-links -->
-
-[issues-shield]: https://img.shields.io/github/issues/Asalz20/bible-verse-scripture-app.svg?style=for-the-badge
-[issues-url]: https://github.com/Asalz20/pectus-plus-backend/issues
-[license-shield]: https://img.shields.io/github/license/Asalz20/bible-verse-scripture-app.svg?style=for-the-badge
-[license-url]: https://github.com/Asalz20/bible-verse-scripture-app/blob/main/LICENSE.txt
-[linkedin-shield]: https://img.shields.io/badge/LinkedIn-0072b1?style=for-the-badge&logo=linkedin&logoColor=fffff
-[linkedin-url]: https://www.linkedin.com/in/alex-salazar-dev/
-
-<!-- [product-screenshot]: ./web/src/images/pectus-plus-blog.png/ -->
-
-[typescript-shield]: https://img.shields.io/badge/TypeScript-20232A?style=for-the-badge&logo=typescript&logoColor=3178C6
-[typescript-url]: https://reactjs.org/
-[react.js]: https://img.shields.io/badge/React-20232A?style=for-the-badge&logo=react&logoColor=61DAFB
-[react-url]: https://reactjs.org/
-[gatsby.js]: https://img.shields.io/badge/Gatsby-BC77C6?style=for-the-badge&logo=gatsby&logoColor=663399
-[gatsby-url]: https://www.gatsbyjs.com/docs/
-[sanity-shield]: https://img.shields.io/badge/-Sanity-FC4C02?style=for-the-badge&logo=stripe&logoColor=white
-[sanity-url]: https://sanity.io/
-[javascript-shield]: https://img.shields.io/badge/-JavaScript-323231?style=for-the-badge&logo=javascript&logoColor=f7df1e
-[javascript-url]: https://developer.mozilla.org/en-US/docs/Web/JavaScript
-[css-shield]: https://img.shields.io/badge/-CSS-AFB3B5?style=for-the-badge&logo=css3&logoColor=1572B6
-[css-url]: https://developer.mozilla.org/en-US/docs/Web/CSS
-[node-shield]: https://img.shields.io/badge/-NodeJs-20232A?style=for-the-badge&logo=node.js&logoColor=339933
-[node-url]: https://nodejs.org/en/docs/
-[express-shield]: https://img.shields.io/badge/-Express-E7E7E7?style=for-the-badge&logo=express&logoColor=339933
-[express-url]: https://expressjs.com/
-[mongo-shield]: https://img.shields.io/badge/-MongoDB-0D0D0D?style=for-the-badge&logo=mongoDB&logoColor=47A248
-[mongo-url]: https://www.mongodb.com/docs/
-[postgres-shield]: https://img.shields.io/badge/-PostgreSQL-0D0D0D?style=for-the-badge&logo=postgreSQL&logoColor=4169E1
-[postgres-url]: https://www.postgresql.org/docs/
