+++ backend/.env
@@ -1,0 +1,5 @@
+DB_USER=alexsalazar
+DB_PASSWORD=password
+DB_HOST=localhost
+DB_PORT=5432
+DB_DATABASE=bvsauth
