--- public/css/style.css
@@ -1,187 +1,0 @@
-.hide {
-  display: none;
-}
-
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-@font-face {
-  * {
-    font-family: Manrope;
-    src: url("https://fonts.google.com/specimen/Manrope");
-  }
-}
-
-body {
-  height: calc(100vh - 0.1px);
-  background-color: rgb(41, 59, 87);
-}
-
-.container {
-  min-height: 90%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-
-.card {
-  position: relative;
-  width: clamp(255px, 95%, 460px);
-  text-align: center;
-  color: rgb(207, 203, 203);
-  background-color: rgb(69, 84, 108);
-  border-radius: 3%;
-  margin: 3rem 0.825rem;
-  padding: 1rem 1rem 2rem 1rem;
-}
-.card .card__passage {
-  padding: 1.25rem 0;
-  font-size: clamp(1.25rem, 3vw, 1.5rem);
-}
-.card .card__passage--highlight {
-  color: aqua;
-}
-.card .card__scripture-container {
-  font-size: clamp(1.25rem, 2vw, 1.5rem);
-  font-weight: 400;
-  max-height: 20.5rem;
-  overflow-y: auto;
-}
-.card .card__separator {
-  width: 100%;
-}
-.card .card__separator img {
-  width: 75%;
-  padding: 1.25rem 0 2rem 0;
-}
-.card .button__container {
-  width: 90%;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 10px;
-  position: absolute;
-}
-.card .button__container .generate__button {
-  padding: 0.5rem;
-  background-color: rgb(0, 255, 255);
-  border-radius: 50%;
-}
-.card .button__container .generate__button #rotate {
-  height: 24px;
-}
-.card .button__container .generate__button:hover {
-  cursor: pointer;
-  background-color: rgb(0, 187, 187);
-}
-.card .button__container .write__button {
-  padding: 0.4rem;
-  background-color: rgb(0, 255, 255);
-  border-radius: 50%;
-}
-.card .button__container .write__button #expand {
-  height: 28px;
-}
-.card .button__container .write__button:hover {
-  cursor: pointer;
-  background-color: rgb(0, 187, 187);
-}
-.card .exit__container {
-  display: flex;
-  width: 95.5%;
-  height: 6.5%;
-  align-items: flex-end;
-  justify-content: flex-end;
-  position: absolute;
-}
-.card .exit__container #close {
-  height: 28px;
-}
-.card .exit__container .exit__button:hover {
-  cursor: pointer;
-  color: red;
-}
-.card .save__container {
-  width: 90%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 10px;
-  position: absolute;
-}
-.card .save__container .save__button {
-  padding: 0.5rem;
-  background-color: rgb(0, 255, 255);
-  border-radius: 50%;
-}
-.card .save__container .save__button #expand {
-  height: 28px;
-}
-.card .save__container .save__button:hover {
-  cursor: pointer;
-  background-color: rgb(0, 187, 187);
-}
-.card .textarea {
-  background-color: rgb(32, 36, 85);
-  color: rgb(207, 203, 203);
-  padding: 0.5rem;
-  font-size: 16px;
-  line-height: 1.1;
-  height: 150px;
-  width: 95%;
-  border-radius: 10px;
-  border: 2px solid transparent;
-  outline: none;
-  transition: all 0.2s;
-  resize: none;
-}
-.card .textarea:hover {
-  cursor: pointer;
-  background-color: rgb(61, 69, 157);
-}
-.card .textarea:focus {
-  cursor: text;
-  color: rgb(156, 156, 156);
-  background-color: rgb(32, 36, 85);
-}
-.card .textarea::-webkit-scrollbar {
-  width: 1rem;
-}
-.card .textarea::-webkit-scrollbar-track {
-  background: rgb(60, 55, 55);
-}
-.card .textarea::-webkit-scrollbar-thumb {
-  background: hsl(100deg, 75%, 10%);
-}
-
-.diceRotate {
-  -webkit-animation: diceRotate 0.5s ease;
-          animation: diceRotate 0.5s ease;
-}
-
-@-webkit-keyframes diceRotate {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
-@keyframes diceRotate {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-.footer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding-top: 20px;
-}/*# sourceMappingURL=style.css.map */
