--- public/scss/_index.scss
@@ -1,7 +1,0 @@
-//theme colors
-$primary: rgb(41, 59, 87, 1);
-$secondary: rgba(203, 137, 76, 1);
-
-.hide {
-    display: none;
-}
