--- public/scss/style.scss
@@ -1,3 +1,0 @@
-@import '../scss/index';
-@import '../scss/card';
-@import '../scss/textbox';
