+++ src/components/Card/StyledCard.ts
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const StyledCard = styled.div`
+  position: relative;
+  width: clamp(255px, 95%, 460px);
+  text-align: center;
+  color: rgb(207, 203, 203);
+  background-color: rgb(69, 84, 108);
+  border-radius: 3%;
+  margin: 3rem 0.825rem;
+  padding: 1rem 1rem 2rem 1rem;
+`;
