+++ src/components/Container/Container.tsx
@@ -1,0 +1,11 @@
+import { StyledContainer } from "./StyledContainer";
+
+interface Props {
+  children: React.ReactNode;
+}
+
+const Container: React.FC<Props> = ({ children }) => {
+  return <StyledContainer>{children}</StyledContainer>;
+};
+
+export default Container;
