+++ src/components/Container/StyledContainer.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const StyledContainer = styled.div`
+  width: 100%;
+  min-height: 90%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+`;
