+++ src/components/CustomInput/StyledCustomInput.ts
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const StyledInput = styled.input`
+  width: 90%;
+`;
