+++ src/components/Footer/StyledFooter.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const StyledFooter = styled.footer`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding-top: 20px;
+`;
