+++ src/components/Header/StyledHeader.ts
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+export const StyledHeader = styled.header`
+  display: flex;
+  min-width: 100vw;
+  & nav {
+    width: 100%;
+  }
+  & ul {
+    display: flex;
+    justify-content: space-around;
+    & :nth-child(1) {
+      width: 60%;
+    }
+  }
+`;
