+++ src/components/Layout.tsx
@@ -1,0 +1,15 @@
+import Header from "./Header/Header";
+import Footer from "./Footer/Footer";
+import Home from "../pages/Home";
+
+function Layout() {
+  return (
+    <>
+      <Header />
+      <Home />
+      <Footer />
+    </>
+  );
+}
+
+export default Layout;
