+++ src/pages/Review.tsx
@@ -1,0 +1,5 @@
+const Review = () => {
+  return <div>Review</div>;
+};
+
+export default Review;
